+++ src/components/burger-constructor/burger-constructor.module.css
@@ -3,7 +3,7 @@
   flex-direction: column;
   padding: 0;
   overflow-y: scroll;
-  height: 800px;
+  height: 600px;
 }
 .burgeritems::-webkit-scrollbar {
   width: 5px;
